requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
bcrypt==4.0.1
PyPDF2==3.0.1
google-generativeai>=0.8.0
langchain==0.2.0
# LangGraph: placeholder package name. Replace with the real package if different.
# If the real package is called `langgraph`, install it. If it's a different
# distribution name, update accordingly.
langgraph==0.2.0
openai==1.0.0
